#40: expose lib method for schema generation
