+++ lib/schema/generate.js
@@ -1,3 +1,5 @@
+const { writeFile } = require('./utils');
+
 function createDefinitions(resources, lookups, additionalProperties = false) {
   const typeMappings = {
     'Edm.String': 'string',
@@ -226,6 +228,16 @@ function generateSchema(metadataJson, additionalProperties) {
   }
 }
 
+async function generateJsonSchema({ metadataReportJson = {}, outputFilePath = '', additionalProperties = false }) {
+  const schema = generateSchema(metadataReportJson, additionalProperties);
+  if (schema && outputFilePath) {
+    const success = await writeFile(outputFilePath, JSON.stringify(schema));
+    if (!success) console.log(`Error writing schema to path ${outputFilePath}`);
+  }
+  return schema;
+}
+
 module.exports = {
-  generateSchema
+  generateSchema,
+  generateJsonSchema
 };
