+++ lib/schema/index.js
@@ -2,48 +2,16 @@
 
 const chalk = require('chalk');
 const { promises: fs } = require('fs');
-const { generateSchema } = require('./generate');
+const { generateSchema, generateJsonSchema } = require('./generate');
 const path = require('path');
 const { validatePayload, isValidDdVersion } = require('./validate');
 const { CURRENT_DATA_DICTIONARY_VERSION } = require('../../common');
 const { readDirectory } = require('../restore-utils');
 const { getReferenceMetadata } = require('reso-certification-etl');
-const { processFiles } = require('./utils');
+const { processFiles, createDirectoryIfNotPresent, writeFile, readFile } = require('./utils');
 
 const OUTPUT_DIR = 'reso-schema-validation-temp';
 const ERROR_REPORT = 'schema-validation-report.json';
-
-const readFile = async filePath => {
-  try {
-    return await fs.readFile(filePath);
-  } catch (err) {
-    console.error(`Could not read file from path '${filePath}'! Error: ${err}`);
-    return null;
-  }
-};
-
-const createDirectoryIfNotPresent = async dirName => {
-  try {
-    await fs.stat(dirName);
-  } catch (error) {
-    try {
-      await fs.mkdir(dirName);
-    } catch (error) {
-      return null;
-    }
-  }
-  return true;
-};
-
-const writeFile = async (path, data) => {
-  try {
-    await fs.writeFile(path, data);
-    return true;
-  } catch (error) {
-    console.log(error);
-    return false;
-  }
-};
 
 const schema = async ({
   metadataPath = '',
@@ -257,5 +225,6 @@ const validatePayloadAndGenerateResults = async ({
 };
 
 module.exports = {
-  schema
+  schema,
+  generateJsonSchema
 };
