+++ lib/schema/utils.js
@@ -42,6 +42,41 @@ const processFile = async ({ filePath, outputPath }) => {
   }
 };
 
+const readFile = async filePath => {
+  try {
+    return await fs.readFile(filePath);
+  } catch (err) {
+    console.error(`Could not read file from path '${filePath}'! Error: ${err}`);
+    return null;
+  }
+};
+
+const createDirectoryIfNotPresent = async dirName => {
+  try {
+    await fs.stat(dirName);
+  } catch (error) {
+    try {
+      await fs.mkdir(dirName);
+    } catch (error) {
+      return null;
+    }
+  }
+  return true;
+};
+
+const writeFile = async (path, data) => {
+  try {
+    await fs.writeFile(path, data);
+    return true;
+  } catch (error) {
+    console.log(error);
+    return false;
+  }
+};
+
 module.exports = {
-  processFiles
+  processFiles,
+  readFile,
+  createDirectoryIfNotPresent,
+  writeFile
 };
